refactor(FavBrowser): move Folder out to own component
