+++ src/components/ui/Sidebar/FavoritesBrowser.tsx
@@ -1,25 +1,17 @@
 import React from 'react';
-import { Link } from 'react-router-dom';
-import {
-  Collapse,
-  IconButton,
-  Typography,
-  List
-} from '@material-tailwind/react';
-import { ChevronRightIcon, FolderIcon } from '@heroicons/react/24/outline';
+import { Collapse, Typography, List } from '@material-tailwind/react';
+import { ChevronRightIcon } from '@heroicons/react/24/outline';
 import { StarIcon as StarFilled } from '@heroicons/react/24/solid';
 
 import ZoneComponent from './Zone';
 import FileSharePathComponent from './FileSharePath';
+import Folder from './Folder';
 import {
   FolderFavorite,
   usePreferencesContext
 } from '@/contexts/PreferencesContext';
-import { useZoneBrowserContext } from '@/contexts/ZoneBrowserContext';
-import { useFileBrowserContext } from '@/contexts/FileBrowserContext';
 import useToggleOpenFavorites from '@/hooks/useToggleOpenFavorites';
 import { FileSharePath, Zone } from '@/shared.types';
-import { makeMapKey } from '@/utils';
 
 type FavoritesBrowserProps = {
   searchQuery: string;
@@ -37,20 +29,8 @@ export default function FavoritesBrowser({
   filteredFolderFavorites
 }: FavoritesBrowserProps) {
   const { openFavorites, toggleOpenFavorites } = useToggleOpenFavorites();
-  const {
-    zoneFavorites,
-    fileSharePathFavorites,
-    folderFavorites,
-    pathPreference,
-    handleFavoriteChange
-  } = usePreferencesContext();
-
-  const { currentDir, fetchAndFormatFilesForDisplay } = useFileBrowserContext();
-  const {
-    setCurrentNavigationZone,
-    currentFileSharePath,
-    setCurrentFileSharePath
-  } = useZoneBrowserContext();
+  const { zoneFavorites, fileSharePathFavorites, folderFavorites } =
+    usePreferencesContext();
 
   const displayZones =
     filteredZoneFavorites.length > 0 || searchQuery.length > 0
@@ -108,70 +88,11 @@ export default function FavoritesBrowser({
 
             {/* Directory favorites */}
             {displayFolders.map(folderFavorite => {
-              const fileSharePath =
-                pathPreference[0] === 'linux_path'
-                  ? folderFavorite.fsp.linux_path
-                  : pathPreference[0] === 'windows_path'
-                    ? folderFavorite.fsp.windows_path
-                    : pathPreference[0] === 'mac_path'
-                      ? folderFavorite.fsp.mac_path
-                      : folderFavorite.fsp.linux_path;
-
-              const mapKey = makeMapKey(
-                'folder',
-                `${folderFavorite.fsp.name}_${folderFavorite.folderPath}`
-              ) as string;
-              const splitPath = folderFavorite.folderPath.split('/');
-              const folderName = splitPath[splitPath.length - 1];
-
               return (
-                <List.Item
-                  key={mapKey}
-                  onClick={() => {
-                    setOpenZones({
-                      all: true,
-                      [folderFavorite.fsp.zone]: true
-                    });
-                    setCurrentNavigationZone(folderFavorite.fsp.zone);
-                    setCurrentFileSharePath(folderFavorite.fsp);
-                    fetchAndFormatFilesForDisplay(
-                      `${folderFavorite.fsp.name}?subpath=${folderFavorite.folderPath}`
-                    );
-                  }}
-                  className={`overflow-x-auto x-short:py-0 flex gap-2 items-center justify-between rounded-none cursor-pointer text-foreground hover:bg-primary-light/30 focus:bg-primary-light/30 ${folderFavorite.fsp === currentFileSharePath && folderFavorite.fsp.name === currentDir ? '!bg-primary-light/30' : '!bg-background'}`}
-                >
-                  <Link
-                    to="/browse"
-                    className="flex flex-col gap-2 x-short:gap-1 !text-foreground hover:!text-black focus:!text-black hover:dark:!text-white focus:dark:!text-white"
-                  >
-                    <div className="flex gap-1 items-center">
-                      <FolderIcon className="icon-small x-short:icon-xsmall" />
-                      <Typography className="text-sm font-medium leading-4 x-short:text-xs">
-                        {folderName}
-                      </Typography>
-                    </div>
-                    <Typography className="text-xs">
-                      {`${fileSharePath}/${folderFavorite.folderPath}`}
-                    </Typography>
-                  </Link>
-                  <div
-                    onClick={e => {
-                      e.stopPropagation();
-                      e.preventDefault();
-                    }}
-                  >
-                    <IconButton
-                      variant="ghost"
-                      isCircular
-                      onClick={(e: React.MouseEvent<HTMLButtonElement>) => {
-                        e.stopPropagation();
-                        handleFavoriteChange(folderFavorite, 'folder');
-                      }}
-                    >
-                      <StarFilled className="icon-small x-short:icon-xsmall mb-[2px]" />
-                    </IconButton>
-                  </div>
-                </List.Item>
+                <Folder
+                  folderFavorite={folderFavorite}
+                  setOpenZones={setOpenZones}
+                />
               );
             })}
           </List>
